completed implementing upto selective ack and increasing the buffer and also implement ring buffer

+++ src/headers/packet.h
@@ -27,7 +27,7 @@ static inline uint64_t ntohll_portable(uint64_t v)
 struct ACKPacket
 {
     uint32_t ack_num; // The sequence number being acknowledged
-    uint16_t bitmap;  // For Selective Repeat (which neighbors are also here)
+    uint64_t bitmap;  // For Selective Repeat (which neighbors are also here) // increase teh bitmap size based on window size
     uint8_t type;     // Always set to 1 (ACK)
     uint32_t crc32;   // Integrity check for the ACK itself
 };
@@ -38,10 +38,10 @@ struct SlimDataPacket
 {
     // --- 14 Bytes Core Header ---
     uint8_t type;
-    uint32_t seq_num;      // 4 bytes (Essential for files > 65MB)
-    uint32_t crc32;        // 4 bytes (Hardware-accelerated)
-    uint16_t data_len;     // 2 bytes (Actual payload size)
-                           
+    uint32_t seq_num;  // 4 bytes (Essential for files > 65MB)
+    uint32_t crc32;    // 4 bytes (Hardware-accelerated)
+    uint16_t data_len; // 2 bytes (Actual payload size)
+
     uint8_t flags;         // 1 byte  (Bit 0: Encrypted, Bit 1-2: Stream ID)
     uint16_t reserved;     // 2 bytes (Padding for 4-byte alignment)
     uint32_t chunk_offset; // 4 bytes (Offset within the file)
